feature 1.支持新增音乐收藏 2.支持取消音乐收藏 3.支持获取音乐收藏列表 4.联调获取媒体数据

+++ app/src/main/java/com/octopus/android/multimedia/room/UserCollection.kt
@@ -9,7 +9,7 @@ import androidx.room.PrimaryKey
 @Entity
 data class UserCollection(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,//id
+    val id: Long = 0,//id
     val path: String,//文件路径
     val time: Long, //收藏时间
 )
